backend/requirements.txt
# YomiGo Backend Dependencies
# ===========================
# Install with: pip install -r requirements.txt

# Web Framework
# FastAPI - Modern, fast web framework for building APIs
# uvicorn - ASGI server to run FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# OCR (Optical Character Recognition)
# manga-ocr - ML model trained specifically for Japanese manga text
# Much more accurate than Tesseract for manga/anime style text
manga-ocr>=0.1.8

# Japanese Text Processing
# fugashi - Python wrapper for MeCab (morphological analyzer)
# unidic-lite - Lightweight dictionary for MeCab (~50MB vs 1GB+ full version)
fugashi>=1.3.0
unidic-lite>=1.0.8

# Dictionary
# jamdict - Python interface to JMdict (Japanese-English dictionary)
# jamdict-data - Pre-packaged JMdict data
jamdict>=0.1.6
jamdict-data>=1.5

# Image Processing
# Pillow - Required for manga-ocr to handle images
Pillow>=10.0.0

# Utilities
# pydantic - Data validation (included with FastAPI, but listed for clarity)
pydantic>=2.0.0
